fixed Solucionado problema de sseguridad de tareas que pueden ser borrado por otros usuarios

+++ app/task/views.py
@@ -1,7 +1,7 @@
 from os import name
 from flask import request
 from marshmallow import ValidationError
-from bson.objectid import ObjectId
+from bson.objectid import ObjectId 
 import datetime
 from flask_jwt_extended import get_jwt_identity, jwt_required
 from pymongo import DESCENDING, ASCENDING
@@ -24,15 +24,12 @@ def createTask():
     data = request.get_json()
     print(current_user)
     data["idUser"] = str(current_user["_id"])
-    print(data)
     try:
         new_task = modelTaskCreate.load(data)
     except ValidationError as err:
         return {"errors": err.messages}, 422
 
-    print("Pase todo esto")
     print(modelTaskCreate.dump(new_task))
-    print("Pase todo esto")
 
     database[TaskBase.DatabaseName].insert_one(modelTaskCreate.dump(new_task))
     return "UserCretated"
@@ -83,14 +80,14 @@ def getTask():
 def deleteTask():
     current_user = get_jwt_identity()
     userId = current_user['_id']
-
     data = request.get_json()
     idTask=data["idTask"]
     _id=ObjectId(idTask)
-
-    dataUser = {"idUser": userId}
-    print(idTask)
-    TaskDatabase.delete_one({"_id":_id})
-    print("Delete was completed")
-
+    task1=TaskDatabase.delete_one({"_id":_id,"idUser":userId})
+    print(task1)
     return {"Task Was Delete":"sd"}
+
+@task_app.route("/task/check", methods=["POST"])
+@jwt_required()
+def checkTask():
+    return "hola"
